Refactor browse_cn into a new class
refs #1763
