+++ lib/orangelight/browse_lists.rb
@@ -18,7 +18,6 @@ module BrowseLists
       # changes for different facet queries
       facet_request = "#{core_url}select?q=*%3A*&fl=id&wt=json&indent=true&defType=edismax&facet.sort=asc&facet.limit=-1&facet.field="
       solr_url = Blacklight.connection_config[:url]
-      # solr_url = 'http://lib-solr1.princeton.edu:8985'
 
       conn = Faraday.new(url: solr_url) do |faraday|
         faraday.options[:open_timeout] = 2000
@@ -53,89 +52,6 @@ module BrowseLists
       end
     end
 
-    # rubocop:disable Metrics/ParameterLists
-    def browse_cn(_sql_command, facet_request, conn, facet_field, _table_name, rows: 50_000)
-      # Collect all the call numbers with facet count > 2
-      # This request takes a minute or so
-      resp = conn.get "#{facet_request}#{facet_field}&facet.mincount=2"
-      req = JSON.parse(resp.body)
-      CSV.open(output_root.join("#{facet_field}.csv"), 'wb') do |csv|
-        mcn = ''
-        multi_cns = {}
-        req['facet_counts']['facet_fields'][facet_field.to_s].each_with_index do |f, i|
-          if i.even?
-            mcn = f
-          else
-            sort_cn = StringFunctions.cn_normalize(mcn)
-            multi_cns[sort_cn] = f
-            csv << [sort_cn, mcn, 'ltr', '', "#{f} titles with this call number", '', '', "?f[#{facet_field}][]=#{CGI.escape(mcn)}", '', 'Multiple locations']
-          end
-        end
-        # Count of all items in the index
-        resp = conn.get "#{core_url}select?q=*%3A*&fl=id&wt=json&indent=true&defType=edismax"
-        num_docs = JSON.parse(resp.body)['response']['numFound']
-        iterations =
-          if (num_docs % rows).zero?
-            num_docs / rows
-          else
-            num_docs / rows + 1
-          end
-
-        start = 0
-        cn_fields = "#{facet_field},title_display,title_vern_display,author_display,author_s,id,pub_created_vern_display,pub_created_display,holdings_1display"
-        iterations.times do
-          # Get the actual items. With 50_000 rows it takes about a minute;
-          # adjust further if there are continued timeout problems
-          cn_request = "#{core_url}select?q=*%3A*&fl=#{cn_fields}&wt=json&indent=true&defType=edismax&facet=false&sort=id%20asc&rows=#{rows}&start=#{start}"
-          resp = conn.get cn_request.to_s
-          req = JSON.parse(resp.body)
-          req['response']['docs'].each do |record|
-            next unless record[facet_field.to_s]
-            record[facet_field.to_s].each_with_index do |cn, _i|
-              sort_cn = StringFunctions.cn_normalize(cn)
-              next if multi_cns.key?(sort_cn)
-              bibid = record['id']
-              title = record['title_display']
-              if record['title_vern_display']
-                title = record['title_vern_display']
-                dir = title.dir
-              else
-                dir = 'ltr' # ltr for non alt script
-              end
-              if record['pub_created_vern_display']
-                date = record['pub_created_vern_display'][0]
-              elsif record['pub_created_display'].present?
-                date = record['pub_created_display'][0]
-              end
-              label = cn
-              if record['author_display']
-                author = record['author_display'][0..1].last
-              elsif record['author_s']
-                author = record['author_s'][0]
-              end
-              if record['holdings_1display']
-                holding_block = JSON.parse(record['holdings_1display'])
-                holding_record = holding_block.select { |_k, h| h['call_number_browse'] == cn }
-                unless holding_record.empty?
-                  if multiple_locations?(holding_record)
-                    location = 'Multiple locations'
-                  else
-                    holding_id = holding_record.keys.first
-                    location = holding_record[holding_id]['location']
-                  end
-                end
-              end
-              holding_id ||= ''
-              location ||= ''
-              csv << [sort_cn, label, dir, '', title, author, date, bibid, holding_id, location]
-            end
-          end
-          start += rows
-        end
-      end
-    end
-    # rubocop:enable Metrics/ParameterLists
-
     def load_facet(sql_command, _facet_request, _conn, _facet_field, table_name)
       system(%(cp /tmp/#{table_name}.sorted /tmp/#{table_name}.sorted.backup))
       system(%(#{sql_command} "TRUNCATE TABLE #{table_name} RESTART IDENTITY;"))
@@ -161,13 +77,5 @@ module BrowseLists
       system(%(#{sql_command} "TRUNCATE TABLE #{table_name} RESTART IDENTITY;"))
       system(%(#{sql_command} "\\copy #{table_name}(sort,label,dir,scheme,title,author,date,bibid,holding_id,location) from '#{sorted_file}' CSV;"))
     end
-
-    private
-
-      # determines if there are multiple locations for the same call number and same bib
-      def multiple_locations?(holdings)
-        locations = holdings.reject { |_k, h| h['library'] == 'Online' }.map { |_k, h| h['location'] }.uniq
-        locations.length > 1
-      end
   end
 end
